+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added grouping by `<optgroup>` tag for `options()` method in `Select` field.
+
 ### Fixed
 - Fixed css for pagination block for `light` theme.